Add files via upload
Objects in Java
Definition:
An object is an instance of a class. It is a basic unit of Object-Oriented Programming (OOP) and represents real-world entities. Objects are created from classes, which act as blueprints for these objects.
Definition:
A class in Java is a blueprint for creating objects. It defines a datatype by bundling data and methods that work on the data into one single unit. A class doesn't occupy any memory until an object is created from it.