Add files via upload
a constructor is a special method that is used to initialize objects. The constructor is called when an object of a class is created. It can be used to set initial values for object attributes or to perform other setup operations. Constructors have the same name as the class and do not have a return type, not even void.

Properties of Constructors
Same Name as Class:

The constructor must have the same name as the class in which it is defined.
No Return Type:

Constructors do not have a return type, not even void.
Automatically Called:

Constructors are automatically called when an object is created using the new keyword.
Types of Constructors:

Default Constructor: A constructor that takes no arguments.
Parameterized Constructor: A constructor that takes one or more arguments.